moved component to better match our folder design philosophy
